Introduce a caching system
Store successful queries in an SQLite database, including taxons,
sequences, and primers. Primer-BLAST job keys are also cached, which
makes it possible to run a query and come back to it later or run
multiple queries at once. A job key without a result is valid for 12
hours. Change some of the existing code to work with the cache